Application for checking codecs selection.
Providing an input that can be used to start a
player that is then used to determine which codec
it would choose with all other things considered
equal. This is achieved by feeding a fake manifest
to AVPlayer that is constructed in such a way that
the player should choose the "preferred" rendition
if it supports the codec.